In this data analysis project, I have chosen a data set of 391 houses in Philadelphia (randomly selected from adds posted between August 2016 to March 2017), which are scraped from website zillow.com.
The project is performed using “R 3.5.1” statistical programming and “R studio” development environment.
The analysis has been divided into five parts for better understanding of the flow.
Part 1 has the information about loading the required packages
Part 2 has the information about loading the data to the R environment,
Part 3 has the process of cleaning the data
Part 4 is an overview of the data with basic descriptive statistics
Part 5 has all the data analysis.